+++ pvinspect/__init__.py
@@ -2,4 +2,4 @@ from . import data, preproc, common
 
 __all__ = ["data", "preproc", "common"]
 
-__version__ = "0.1.5"
+__version__ = "0.1.6"